able to run a build of the frontend
